+++ tests/_async/test_storages.py
@@ -3,12 +3,10 @@ import os
 from pathlib import Path
 
 import anysqlite
-import boto3
 import pytest
 from httpcore import Request, Response
-from moto import mock_s3
 
-from hishel import AsyncFileStorage, AsyncInMemoryStorage, AsyncRedisStorage, AsyncS3Storage, AsyncSQLiteStorage
+from hishel import AsyncFileStorage, AsyncInMemoryStorage, AsyncRedisStorage, AsyncSQLiteStorage
 from hishel._serializers import Metadata
 from hishel._utils import asleep, generate_key
 
@@ -25,50 +23,9 @@ async def is_redis_down() -> bool:
         return True
 
 
-@pytest.fixture(scope="function")
-def aws_credentials():
-    """Mocked AWS Credentials for moto."""
-
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-    os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
-
-
-@pytest.fixture(scope="function")
-def s3(aws_credentials):
-    with mock_s3():
-        yield boto3.client("s3", region_name="us-east-1")
-
-
 @pytest.mark.anyio
 async def test_filestorage(use_temp_dir):
     storage = AsyncFileStorage()
-
-    request = Request(b"GET", "https://example.com")
-
-    key = generate_key(request)
-
-    response = Response(200, headers=[], content=b"test")
-    await response.aread()
-
-    await storage.store(key, response=response, request=request, metadata=dummy_metadata)
-
-    stored_data = await storage.retrieve(key)
-    assert stored_data is not None
-    stored_response, stored_request, metadata = stored_data
-    stored_response.read()
-    assert isinstance(stored_response, Response)
-    assert stored_response.status == 200
-    assert stored_response.headers == []
-    assert stored_response.content == b"test"
-
-
-@pytest.mark.anyio
-async def test_s3storage(use_temp_dir, s3):
-    boto3.client("s3").create_bucket(Bucket="testBucket")
-    storage = AsyncS3Storage(bucket_name="testBucket")
 
     request = Request(b"GET", "https://example.com")
 
@@ -182,29 +139,6 @@ async def test_filestorage_expired(use_temp_dir, anyio_backend):
 
 
 @pytest.mark.parametrize("anyio_backend", ["asyncio"])
-async def test_s3storage_expired(use_temp_dir, s3, anyio_backend):
-    boto3.client("s3").create_bucket(Bucket="testBucket")
-    storage = AsyncS3Storage(bucket_name="testBucket", ttl=3)
-
-    first_request = Request(b"GET", "https://example.com")
-    second_request = Request(b"GET", "https://anotherexample.com")
-
-    first_key = generate_key(first_request)
-    second_key = generate_key(second_request)
-
-    response = Response(200, headers=[], content=b"test")
-    await response.aread()
-
-    await storage.store(first_key, response=response, request=first_request, metadata=dummy_metadata)
-    assert await storage.retrieve(first_key) is not None
-
-    await asleep(3)
-    await storage.store(second_key, response=response, request=second_request, metadata=dummy_metadata)
-
-    assert await storage.retrieve(first_key) is None
-
-
-@pytest.mark.parametrize("anyio_backend", ["asyncio"])
 async def test_filestorage_timer(use_temp_dir, anyio_backend):
     storage = AsyncFileStorage(ttl=0.2, check_ttl_every=0.2)
 
@@ -231,6 +165,36 @@ async def test_filestorage_timer(use_temp_dir, anyio_backend):
 
 
 @pytest.mark.parametrize("anyio_backend", ["asyncio"])
+async def test_filestorage_ttl_after_hits(use_temp_dir, anyio_backend):
+    storage = AsyncFileStorage(ttl=0.2, check_ttl_every=0.2)
+
+    request = Request(b"GET", "https://example.com")
+
+    key = generate_key(request)
+
+    response = Response(200, headers=[], content=b"test")
+    await response.aread()
+
+    # Storing
+    await storage.store(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.08 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.16 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.24 second
+    await asleep(0.08)
+    assert await storage.retrieve(key) is None
+
+
+@pytest.mark.parametrize("anyio_backend", ["asyncio"])
 async def test_redisstorage_expired(anyio_backend):
     if await is_redis_down():  # pragma: no cover
         pytest.fail("Redis server was not found")
@@ -251,6 +215,36 @@ async def test_redisstorage_expired(anyio_backend):
     await storage.store(second_key, response=response, request=second_request, metadata=dummy_metadata)
 
     assert await storage.retrieve(first_key) is None
+
+
+@pytest.mark.parametrize("anyio_backend", ["asyncio"])
+async def test_redis_ttl_after_hits(use_temp_dir, anyio_backend):
+    storage = AsyncRedisStorage(ttl=0.2)
+
+    request = Request(b"GET", "https://example.com")
+
+    key = generate_key(request)
+
+    response = Response(200, headers=[], content=b"test")
+    await response.aread()
+
+    # Storing
+    await storage.store(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.08 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.16 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.24 second
+    await asleep(0.08)
+    assert await storage.retrieve(key) is None
 
 
 @pytest.mark.parametrize("anyio_backend", ["asyncio"])
@@ -275,6 +269,36 @@ async def test_sqlite_expired(anyio_backend):
 
 
 @pytest.mark.parametrize("anyio_backend", ["asyncio"])
+async def test_sqlite_ttl_after_hits(use_temp_dir, anyio_backend):
+    storage = AsyncSQLiteStorage(ttl=0.2)
+
+    request = Request(b"GET", "https://example.com")
+
+    key = generate_key(request)
+
+    response = Response(200, headers=[], content=b"test")
+    await response.aread()
+
+    # Storing
+    await storage.store(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.08 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.16 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.24 second
+    await asleep(0.08)
+    assert await storage.retrieve(key) is None
+
+
+@pytest.mark.parametrize("anyio_backend", ["asyncio"])
 async def test_inmemory_expired(anyio_backend):
     storage = AsyncInMemoryStorage(ttl=0.1)
     first_request = Request(b"GET", "https://example.com")
@@ -293,6 +317,36 @@ async def test_inmemory_expired(anyio_backend):
     await storage.store(second_key, response=response, request=second_request, metadata=dummy_metadata)
 
     assert await storage.retrieve(first_key) is None
+
+
+@pytest.mark.parametrize("anyio_backend", ["asyncio"])
+async def test_inmemory_ttl_after_hits(use_temp_dir, anyio_backend):
+    storage = AsyncInMemoryStorage(ttl=0.2)
+
+    request = Request(b"GET", "https://example.com")
+
+    key = generate_key(request)
+
+    response = Response(200, headers=[], content=b"test")
+    await response.aread()
+
+    # Storing
+    await storage.store(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.08 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.16 second
+    await asleep(0.08)
+    await storage.update_metadata(key, response=response, request=request, metadata=dummy_metadata)
+    assert await storage.retrieve(key) is not None
+
+    # Retrieving after 0.24 second
+    await asleep(0.08)
+    assert await storage.retrieve(key) is None
 
 
 @pytest.mark.anyio
@@ -316,42 +370,3 @@ async def test_filestorage_empty_file_exception(use_temp_dir):
         file.truncate(0)
     assert os.path.getsize(filedir) == 0
     assert await storage.retrieve(key) is None
-
-
-@pytest.mark.anyio
-async def test_s3storage_timer(use_temp_dir, s3):
-    boto3.client("s3").create_bucket(Bucket="testBucket")
-    storage = AsyncS3Storage(bucket_name="testBucket", ttl=5, check_ttl_every=5)
-
-    first_request = Request(b"GET", "https://example.com")
-    second_request = Request(b"GET", "https://anotherexample.com")
-
-    first_key = generate_key(first_request)
-    second_key = generate_key(second_request)
-
-    response = Response(200, headers=[], content=b"test")
-    await response.aread()
-
-    await storage.store(first_key, response=response, request=first_request, metadata=dummy_metadata)
-    assert await storage.retrieve(first_key) is not None
-    await asleep(3)
-    assert await storage.retrieve(first_key) is not None
-    await storage.store(second_key, response=response, request=second_request, metadata=dummy_metadata)
-    assert await storage.retrieve(second_key) is not None
-    await asleep(2)
-    assert await storage.retrieve(first_key) is None
-    assert await storage.retrieve(second_key) is not None
-    await asleep(3)
-    assert await storage.retrieve(second_key) is None
-
-
-@pytest.mark.anyio
-async def test_s3storage_key_error(use_temp_dir, s3):
-    """Triggers `NoSuchKey` error."""
-
-    boto3.client("s3").create_bucket(Bucket="testBucket")
-    storage = AsyncS3Storage(bucket_name="testBucket", ttl=60)
-    first_request = Request(b"GET", "https://example.com")
-    first_key = generate_key(first_request)
-
-    assert await storage.retrieve(first_key) is None